backend/requirements.txt
fastapi
uvicorn
pydantic
mysql-connector-python
pymysql
passlib[bcrypt]
pyjwt
python-dotenv
email-validator


twilio 